[YouTube] Extract uploader URL for videos with multiple uploaders
Fixes an issue, where for videos with multiple uploaders/collaborators, the
uploader URL could not be extracted, leading to the entire video failing to be
extracted. This is fixed by extracting the URL of the first collaborator, which
appears to be the main channel where the video is actually displayed.

Closes: https://github.com/TeamNewPipe/NewPipeExtractor/issues/1333